CHORE: rename src directory to project name
