+++ fsm_tester/adapters/__init__.py
@@ -1,0 +1,9 @@
+from fsm_tester.adapters.base_adapter import BaseAdapter
+from fsm_tester.adapters.transitions_adapter import TransitionsAdapter
+from fsm_tester.adapters.adapter_factory import AdapterFactory
+
+__all__ = [
+    'AdapterFactory',
+    'BaseAdapter',
+    'TransitionsAdapter',
+]
