+++ fsm_tester/components/rich_console.py
