+++ fsm_tester/entities/__init__.py
@@ -1,0 +1,6 @@
+from fsm_tester.entities.fsm_protocol import FSMProtocol
+from fsm_tester.entities.fsm_state import FSMState
+from fsm_tester.entities.fsm_transition import FSMTransition
+from fsm_tester.entities.testcase import TestCase
+
+__all__ = ['FSMProtocol', 'FSMState', 'FSMTransition', 'TestCase']
