--- src/adapters/__init__.py
@@ -1,9 +1,0 @@
-from src.adapters.base_adapter import BaseAdapter
-from src.adapters.transitions_adapter import TransitionsAdapter
-from src.adapters.adapter_factory import AdapterFactory
-
-__all__ = [
-    'AdapterFactory',
-    'BaseAdapter',
-    'TransitionsAdapter',
-]
