--- src/entities/__init__.py
@@ -1,6 +1,0 @@
-from src.entities.fsm_protocol import FSMProtocol
-from src.entities.fsm_state import FSMState
-from src.entities.fsm_transition import FSMTransition
-from src.entities.testcase import TestCase
-
-__all__ = ['FSMProtocol', 'FSMState', 'FSMTransition', 'TestCase']
