+++ tests/test_unreachable_machine.py
@@ -1,6 +1,6 @@
 import pytest  # noqa
 from machines.defective.unreachable import UnreachableMachine
-from src.fsm_tester import FSMTester
+from fsm_tester.fsm_tester import FSMTester
 
 
 @pytest.fixture
